+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [0.0.21] - 2026-07-28
 
+## Added
+
+- ✨(conversation) summarize messages
+
 ### Changed
 
 - ⬆️(dependencies) update pypdf and next
@@ -120,7 +124,6 @@ and this project adheres to
 - 🐛(fix) add prevent_url_hallucination instruction to ConversationAgent
 - ✨(projects) handle project files for RAG search
 - ✨(banner) configurable banner with level, title, content and start/end
-- ✨(conversation) summarize messages
 
 ### Changed